Add files via upload
Data Cleaning Project

This project demonstrates the data cleaning process, showcasing the transformation of a data table across several sheets. The objective is to provide a clear and structured approach to cleaning and preparing data for analysis. Each step in the process is meticulously documented, ensuring transparency and reproducibility.

Data Source
The dataset used in this project is sourced from Kaggle: https://www.kaggle.com/datasets/imakash3011/customer-personality-analysis/data